feature/cmake-vcpkg-toolchain: Separate spark and databricks tests
