+++ tests/databricks/serverless_databricks_cluster.cpp
@@ -5,7 +5,7 @@
 #include "session.h"
 #include "config.h"
 #include "dataframe.h"
-#include "env_loader.h"
+#include "../util/env_loader.h"
 
 #include <iostream>
 #include <sstream>
