+++ tests/spark/dataframe.cpp
@@ -9,7 +9,7 @@
 #include "dataframe.h"
 #include "functions.h"
 #include "types.h"
-#include "spark_integration.h"
+#include "spark_fixture.h"
 
 using namespace spark::sql::functions;
 using namespace spark::sql::types;
