+++ tests/spark/dataframe_columns.cpp
