+++ tests/spark/expressions.cpp
@@ -6,7 +6,7 @@
 #include "config.h"
 #include "functions.h"
 #include "types.h"
-#include "spark_integration.h"
+#include "spark_fixture.h"
 
 using namespace spark::sql::functions;
 using namespace spark::sql::types;
