+++ tests/spark/spark_fixture.h
